Added ProblemDetailsException with new ProblemDetails extensions

+++ src/services/projects/BldIt.Projects.Api/Controllers/ProjectController.cs
@@ -3,11 +3,14 @@ using BldIt.Api.Shared.Abstractions;
 using BldIt.Api.Shared.Config;
 using BldIt.Api.Shared.Exceptions;
 using BldIt.Api.Shared.Responses;
-using BldIt.Api.Shared.Services;
+using BldIt.Api.Shared.Responses.Problems;
 using BldIt.Api.Shared.Services.Errors;
-using BldIt.Projects.Contracts.Dtos;
+using BldIt.Api.Shared.Services.Uri;
+using BldIt.Projects.Contracts.Contracts;
+using BldIt.Projects.Core.Dtos;
 using BldIt.Projects.Core.Models;
 using BldIt.Projects.Core.Repos;
+using MassTransit;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -20,15 +23,21 @@ namespace BldIt.Projects.Service.Controllers
         private readonly IProjectRepo _projectsRepository;
         private readonly UriService _uriService;
         private readonly BldItWorkspacePathConfig _bldItWorkspacePathConfig;
+        private readonly IPublishEndpoint _publishEndpoint;
+        private readonly ILogger<ProjectController> _logger;
 
         public ProjectController(
             IProjectRepo projectsRepository,
             UriService uriService, 
-            BldItWorkspacePathConfig bldItWorkspacePathConfig)
+            BldItWorkspacePathConfig bldItWorkspacePathConfig, 
+            IPublishEndpoint publishEndpoint, 
+            ILogger<ProjectController> logger)
         {
             _projectsRepository = projectsRepository;
             _uriService = uriService;
             _bldItWorkspacePathConfig = bldItWorkspacePathConfig;
+            _publishEndpoint = publishEndpoint;
+            _logger = logger;
         }
         
         [HttpGet(Routes.Projects.GetAll)]
@@ -55,12 +64,21 @@ namespace BldIt.Projects.Service.Controllers
         public async Task<IActionResult> Get([FromRoute] Guid projectId)
         {
             var project = await _projectsRepository.GetAsync(projectId);
-            if (project == null) 
-                throw new DomainNotFoundException($"Project with id '{projectId}' was not found");
+            if (project == null)
+            {
+                throw new ProblemDetailsException(new InstanceNotFound(
+                    $"Project with id '{projectId}' was not found",
+                    _uriService.GetProjectUri(projectId.ToString()).AbsolutePath,
+                    _uriService));
+            }
 
+            //If user does not own the project, log a 403, BUT return a 404 (for security purposes...like GitHub does)
             if (!await _projectsRepository.IsUserOwnerOfProject(Guid.Parse(UserId), projectId))
-                throw new DomainForbiddenException("You are not the owner of this project");
+                throw Log403Throw404(projectId.ToString());
             
+            //TODO: Check if user is a collaborator
+            
+            //Otherwise, return it
             return Ok(project);
         }
         
@@ -68,11 +86,16 @@ namespace BldIt.Projects.Service.Controllers
         public async Task<IActionResult> GetWithQuery([FromQuery] string projectName)
         {
             var project = await _projectsRepository.GetByNameAsync(projectName);
-            if (project == null) 
-                throw new DomainNotFoundException($"Project with name '{projectName}' was not found");
+            if (project == null)
+            {
+                throw new ProblemDetailsException(new InstanceNotFound(
+                    $"Project with id '{projectName}' was not found",
+                    HttpContext.Request.Path.ToString(),
+                    _uriService));
+            }
             
             if (!await _projectsRepository.IsUserOwnerOfProject(Guid.Parse(UserId), project.Id))
-                throw new DomainForbiddenException("You are not the owner of this project");
+                throw Log403Throw404(projectName);
             
             return Ok(project);
         }
@@ -94,14 +117,10 @@ namespace BldIt.Projects.Service.Controllers
 
             if (existingProject != null)
             {
-                throw new ProblemDetailsException(new ProblemDetails
-                {
-                    Detail = "All projects have unique names, there cannot be a project with the same name",
-                    Instance = Routes.Projects.Post,
-                    Status = StatusCodes.Status400BadRequest,
-                    Title = ErrorTypeMessages.ExistingInstance,
-                    Type = _uriService.GetDocsErrorTypeUri(ErrorTypeMessages.ExistingInstance).ToString()
-                });
+                throw new ProblemDetailsException(new ExistingInstance(
+                    $"Project with name '{projectToCreate.ProjectName}' already exists",
+                    Routes.Projects.Post,
+                    _uriService));
             }
 
             var projPath = Path.Combine(_bldItWorkspacePathConfig.ProjectsPath, projectToCreate.ProjectName);
@@ -117,6 +136,9 @@ namespace BldIt.Projects.Service.Controllers
             
             await _projectsRepository.CreateAsync(project);
             
+            //Publish project created event
+            await _publishEndpoint.Publish(new ProjectCreated(project.Id, project.ProjectWorkspacePath));
+            
             var locationUri = _uriService.GetProjectUri(project.ProjectName);
             return Created(locationUri, project);
         }
@@ -125,16 +147,17 @@ namespace BldIt.Projects.Service.Controllers
         public async Task<IActionResult> Delete([FromRoute] Guid projectId)
         {
             var project = await _projectsRepository.GetAsync(projectId);
-            if (project == null) 
-                throw new ProblemDetailsException(new ProblemDetails
-                {
-                    Detail = $"Project with id '{projectId}' was not found",
-                    Instance = _uriService.GetProjectUri(projectId.ToString()).AbsolutePath,
-                    Status = StatusCodes.Status404NotFound,
-                    Title = ErrorTypeMessages.InstanceNotFound,
-                    Type = _uriService.GetDocsErrorTypeUri(ErrorTypeMessages.InstanceNotFound).ToString()
-                });
+            if (project == null)
+            {
+                throw new ProblemDetailsException(new InstanceNotFound(
+                    $"Project with id '{projectId}' was not found",
+                    _uriService.GetProjectUri(projectId.ToString()).AbsolutePath,
+                    _uriService));
+            }
             
+            if (!await _projectsRepository.IsUserOwnerOfProject(Guid.Parse(UserId), project.Id))
+                throw Log403Throw404(projectId.ToString());
+
             if(Directory.Exists(project.ProjectWorkspacePath))
                 Directory.Delete(project.ProjectWorkspacePath, true);
 
@@ -147,6 +170,28 @@ namespace BldIt.Projects.Service.Controllers
         {
             if(!Directory.Exists(projectWorkspacePath))
                 Directory.CreateDirectory(projectWorkspacePath);
+        }
+
+        /// <summary>
+        /// Returns ProblemDetailsException with InstanceNotFound
+        /// but logs a 403 for the user trying to access a project that does not own
+        /// </summary>
+        /// <param name="project">Project Id trying to be accessed</param>
+        /// <returns>ProblemDetailsException with InstanceNotFound</returns>
+        /// <remarks>
+        /// throw this method as if it was a ProblemDetailsException.
+        /// https://softwareengineering.stackexchange.com/questions/207232/are-there-legitimate-reasons-for-returning-exception-objects-instead-of-throwing
+        /// </remarks>
+        private ProblemDetailsException Log403Throw404(string project)
+        {
+            _logger.LogWarning(
+                "User \'{UserId}\' tried to access unowned project \'{Project}\' (403)", 
+                UserId, project);
+            
+            return new ProblemDetailsException(new InstanceNotFound(
+                $"Project with '{project}' was not found",
+                _uriService.GetProjectUri(project).AbsolutePath,
+                _uriService));
         }
     }
 }
